fix(types): break circular dependency between identity.ts and index.ts
Extract AttestationPayload to attestation.ts so identity.ts can import
it without creating a barrel-file cycle through index.ts.

Fixes CI Quality Guard circular dependency check failure.
