+++ packages/types/src/attestation.ts
@@ -1,0 +1,7 @@
+/** Device attestation payload sent during identity verification. */
+export interface AttestationPayload {
+  platform: 'ios' | 'android' | 'web';
+  integrityToken: string;
+  deviceKey: string;
+  nonce: string;
+}
